chore(layout): bring the Rust layout gate back to green
`Rust Quality` — and therefore `PR CI Gate` — is failing on `main` itself
(61d25fe21), so no PR branched from it can go green. Three files trip
`scripts/ci/check-openhuman-rust-layout.mjs`:

  - session/turn/context.rs: inline test module
  - tools/impl/filesystem/git_operations.rs: 949 lines (limit 750)
  - tools/impl/filesystem/git_operations_tests.rs: 929 lines (limit 750)

Pure code movement — no behaviour change, nothing renamed, no test dropped.

`context.rs` keeps its four tool-policy-boundary tests, moved to the sibling
`context_tests.rs` the gate asks for.

`git_operations.rs` sheds two cohesive units rather than being cut at an
arbitrary line: the ambient-config hardening (the allow-list, the neutralised
keys and the helpers that build a hardened `git` invocation) becomes
`git_operations_config.rs`, and the result formatters become
`git_operations_render.rs`. Reviewing the config policy is easier when it is
not interleaved with the command implementations.

The tests split along the same seam, so the #5494 hardening tests sit beside
the module they exercise. `git_operations_config_tests.rs` attaches to
`git_operations` rather than to the config module, so `use super::*` still
resolves `GitOperationsTool` exactly as before; the fixtures stay in
`git_operations_tests.rs` and are shared via `pub(super)` rather than
duplicated, since both modules are children of `git_operations`.

Test-count parity checked against main: git_operations 42 = 30 + 12, and
context.rs 4 = 4.

Note for reviewers: `cargo test --lib git_operations` reports 35 passed /
7 failed **both before and after this commit**, with an identical set of
failing test names. Those seven are the #5494 config-hardening tests failing
in a local sandbox ("Not in a git repository"); they are pre-existing and
untouched by this change, not a regression introduced here.
